Move all of the Domain-related stuff into global noa src to use in pms
