--- test/cfd/ConfigTagPermissive.hh
@@ -1,8 +1,0 @@
-#pragma once
-
-template <typename CellTopology> struct ConfigTagPermissive {};
-
-namespace noa::TNL::Meshes::BuildConfigTags {
-	template <typename CellTopology>
-	struct MeshCellTopologyTag<ConfigTagPermissive<CellTopology>, CellTopology> { enum { enabled = true }; };
-}
